feat(audio): complete bus routing wireup — voices route to buses, sends, returns
P0 bus routing system fully wired:
- HookGraphEngine.process_into_buses(): voices route to assigned OutputBus via BusBuffers
- render_voices_to_buses(): thread-local scratch buffers (zero audio-thread allocation)
- BusSendNode: rewritten with one-pole smoother for click-free level changes
- BusReturnNode: NEW node with smoothed return level, mute state, 3 unit tests
- SetBusVolume command wired in drain_commands()
- playback.rs: calls process_into_buses() instead of legacy process()
- AudioBuffer::with_constant() + NodeContext::default() for test convenience
- OrbMixer architecture doc (3 interaction levels, 4 visual layers)
- MASTER_TODO updated: bus routing marked COMPLETE

Build: cargo build --release clean, 53 tests pass

+++ crates/rf-spatial/src/hoa/transform.rs
@@ -528,7 +528,7 @@ fn wigner_d_real(l: usize, beta: f32) -> Vec<Vec<f32>> {
     let j = l as i64;
 
     // Pre-compute factorials (up to 2*l which is max 14 for order 7)
-    let max_fact = (2 * l + 1) as usize;
+    let max_fact = 2 * l + 1;
     let mut fact = vec![1.0_f64; max_fact + 1];
     for i in 1..=max_fact {
         fact[i] = fact[i - 1] * i as f64;
